backend/config/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9
python-decouple==3.8
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.0
google-api-python-client==2.108.0
pillow==10.1.0
django-extensions==3.2.3
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0